+++ src/Router.tsx
@@ -1,17 +1,17 @@
 import { createBrowserRouter } from "react-router-dom";
 
 export const router = createBrowserRouter([
-  //   {
-  //     path: "/1",
-  //     element: <Page1 />,
-  //   },
-  //   {
-  //     path: "/2",
-  //     element: <Page2 />,
-  //   },
-  //   {
-  //     path: "/3",
-  //     element: <Page3 />,
-  //   },
-  //   ...
+  // {
+  //   path: "/1",
+  //   element: <Page1 />,
+  // },
+  // {
+  //   path: "/2",
+  //   element: <Page2 />,
+  // },
+  // {
+  //   path: "/3",
+  //   element: <Page3 />,
+  // },
+  // ...
 ]);
