+++ src/assets/index.ts
@@ -3,7 +3,4 @@ import IcLoud from "./svg/ic_loud.svg?react";
 import IcSym from "./svg/ic_sym.svg?react";
 import IcTopBar from "./svg/ic_top_bar.svg?react";
 
-export { IcLogoSim };
-export { IcLoud };
-export { IcSym };
-export { IcTopBar };
+export { IcLogoSim, IcLoud, IcSym, IcTopBar };
